Refactor ResourceManager to IAM
