+++ ScoutSuite/output/data/html/partials/gcp/services.iam.projects.id.bindings.html
@@ -1,16 +1,16 @@
 
 <!-- Cloud Resources Manager bindings' partial -->
-<script id="services.cloudresourcemanager.projects.id.bindings.partial" type="text/x-handlebars-template">
+<script id="services.iam.projects.id.bindings.partial" type="text/x-handlebars-template">
     <div id="resource-name" class="list-group-item active">
         <h4 class="list-group-item-heading">{{name}}</h4>
     </div>
     <div class="list-group-item">
         <h4 class="list-group-item-heading">Information</h4>
-        <div class="list-group-item-text item-margin">Role Name: <span id="cloudresourcemanager.projects.{{@../key}}.bindings.{{@key}}.name"><samp>{{name}}</samp></span></div>
+        <div class="list-group-item-text item-margin">Role Name: <span id="iam.projects.{{@../key}}.bindings.{{@key}}.name"><samp>{{name}}</samp></span></div>
     </div>
     <div class="list-group-item">
         <h4 class="list-group-item-heading">Bindings</h4>
-        <h5  id="cloudresourcemanager.projects.{{@../key}}.bindings.{{@key}}.users" class="list-group-item-heading">Attached Users:</h5>
+        <h5  id="iam.projects.{{@../key}}.bindings.{{@key}}.users" class="list-group-item-heading">Attached Users:</h5>
         <div class="accordion-inner">
             <ul>
                 {{#each members.users}}
@@ -20,7 +20,7 @@
                 {{/each}}
             </ul>
         </div>
-        <h5  id="cloudresourcemanager.projects.{{@../key}}.bindings.{{@key}}.groups" class="list-group-item-heading">Attached Groups:</h5>
+        <h5  id="iam.projects.{{@../key}}.bindings.{{@key}}.groups" class="list-group-item-heading">Attached Groups:</h5>
         <div class="accordion-inner">
             <ul>
                 {{#each members.groups}}
@@ -30,7 +30,7 @@
                 {{/each}}
             </ul>
         </div>
-        <h5  id="cloudresourcemanager.projects.{{@../key}}.bindings.{{@key}}.service_accounts" class="list-group-item-heading">Attached Service Accounts:</h5>
+        <h5  id="iam.projects.{{@../key}}.bindings.{{@key}}.service_accounts" class="list-group-item-heading">Attached Service Accounts:</h5>
         <div class="accordion-inner">
             <ul>
                 {{#each members.service_accounts}}
@@ -44,13 +44,13 @@
 </script>
 
 <script>
-    Handlebars.registerPartial("services.cloudresourcemanager.projects.id.bindings", $("#services\\.cloudresourcemanager\\.projects\\.id\\.bindings\\.partial").html());
+    Handlebars.registerPartial("services.iam.projects.id.bindings", $("#services\\.iam\\.projects\\.id\\.bindings\\.partial").html());
 </script>
 
-<!-- Single cloudresourcemanager binding template -->
-<script id="single_cloudresourcemanager_binding-template" type="text/x-handlebars-template">
-    {{> modal-template template='services.cloudresourcemanager.projects.id.bindings'}}
+<!-- Single iam binding template -->
+<script id="single_iam_binding-template" type="text/x-handlebars-template">
+    {{> modal-template template='services.iam.projects.id.bindings'}}
 </script>
 <script>
-    var single_cloudresourcemanager_binding_template = Handlebars.compile($("#single_cloudresourcemanager_binding-template").html());
+    var single_iam_binding_template = Handlebars.compile($("#single_iam_binding-template").html());
 </script>
