+++ ScoutSuite/output/data/html/partials/gcp/services.iam.projects.id.groups.html
@@ -1,0 +1,35 @@
+
+<!-- Cloud Resources Manager groups' partial -->
+<script id="services.iam.projects.id.groups.partial" type="text/x-handlebars-template">
+    <div id="resource-name" class="list-group-item active">
+        <h4 class="list-group-item-heading">{{name}}</h4>
+    </div>
+    <div class="list-group-item">
+        <h4 class="list-group-item-heading">Information</h4>
+        <div class="list-group-item-text item-margin">Group: <span id="iam.projects.{{@../key}}.groups.{{@key}}.name">{{name}}</span></div>
+    </div>
+    <div class="list-group-item">
+        <h4 class="list-group-item-heading">Bindings:</h4>
+        <div class="accordion-inner">
+            <ul>
+                {{#each roles}}
+                <li><samp>{{this}}</samp></li>
+                {{else}}
+                <li><samp>None</samp></li>
+                {{/each}}
+            </ul>
+        </div>
+    </div>
+</script>
+
+<script>
+    Handlebars.registerPartial("services.iam.projects.id.groups", $("#services\\.iam\\.projects\\.id\\.groups\\.partial").html());
+</script>
+
+<!-- Single iam group template -->
+<script id="single_iam_group-template" type="text/x-handlebars-template">
+    {{> modal-template template='services.iam.projects.id.groups'}}
+</script>
+<script>
+    var single_iam_group_template = Handlebars.compile($("#single_iam_group-template").html());
+</script>
