+++ ScoutSuite/providers/gcp/resources/iam/groups.py
