+++ ScoutSuite/providers/gcp/resources/iam/service_account_bindings.py
