+++ ScoutSuite/providers/gcp/resources/iam/service_accounts.py
@@ -1,6 +1,6 @@
 from ScoutSuite.providers.gcp.facade.base import GCPFacade
 from ScoutSuite.providers.gcp.resources.base import GCPCompositeResources
-from ScoutSuite.providers.gcp.resources.iam.bindings import ServiceAccountBindings
+from ScoutSuite.providers.gcp.resources.iam.service_account_bindings import ServiceAccountBindings
 from ScoutSuite.providers.gcp.resources.iam.keys import Keys
 import re
