+++ ScoutSuite/providers/gcp/rules/findings/iam-role-assigned-to-user.json
@@ -6,11 +6,11 @@
         "https://cloud.google.com/iam/docs/using-iam-securely"
     ],
     "dashboard_name": "Bindings",
-    "path": "cloudresourcemanager.projects.id.bindings.id",
+    "path": "iam.projects.id.bindings.id",
     "conditions": [
         "and",
         [
-            "cloudresourcemanager.projects.id.bindings.id.members.users",
+            "iam.projects.id.bindings.id.members.users",
             "notEmpty",
             ""
         ]
