+++ ScoutSuite/providers/gcp/rules/findings/iam-user-has-sa-user-role.json
@@ -14,11 +14,11 @@
         "https://cloud.google.com/iam/docs/using-iam-securely"
     ],
     "dashboard_name": "Bindings",
-    "path": "cloudresourcemanager.projects.id.bindings.id",
+    "path": "iam.projects.id.bindings.id",
     "conditions": [
         "and",
         [
-            "cloudresourcemanager.projects.id.bindings.id.name",
+            "iam.projects.id.bindings.id.name",
             "containAtLeastOneOf",
             [
                 "iam.serviceAccountUser"
